Fixed some bugs and created and implemented logic to store sales in the database

+++ src/main/java/com/br/BookStoreAPI/models/DTOs/bookDTOs/BookDetailsResponseDTO.java
@@ -10,7 +10,7 @@ public record BookDetailsResponseDTO(
         int pageNumbers,
         String author,
         String category,
-        float price,
+        Double price,
         LocalDateTime creationDate,
         LocalDateTime modificationDate
